[process] Refactor user stack/args functions into portable code
