+++ config/config_mm.h
@@ -43,7 +43,6 @@
 #define USER_MMAP_BEGIN               MAX_BRK /* +1 GB (virtual memory) */
 #define USER_MMAP_MIN_SZ            (16 * MB)
 #define USER_MMAP_MAX_SZ          (1024 * MB)
-#define USERMODE_STACK_ALIGN              16u
 
 #define USERMODE_STACK_MAX \
    ((USERMODE_VADDR_END - 1) & ALIGNED_MASK(USERMODE_STACK_ALIGN))
