+++ kernel/arch/riscv/process_riscv.c
@@ -38,12 +38,6 @@ STATIC_ASSERT(
 
 STATIC_ASSERT(sizeof(struct task_and_process) <= 2048);
 
-void task_info_reset_kernel_stack(struct task *ti)
-{
-
-   NOT_IMPLEMENTED();
-}
-
 void setup_pause_trampoline(regs_t *r)
 {
    NOT_IMPLEMENTED();
@@ -57,6 +51,22 @@ void setup_pause_trampoline(regs_t *r)
          - sizeof(regs_t)               /* regs */      \
       ) % USERMODE_STACK_ALIGN                          \
    )
+
+int
+push_args_on_user_stack(regs_t *r,
+                        const char *const *argv,
+                        u32 argc,
+                        const char *const *env,
+                        u32 envc)
+{
+   NOT_IMPLEMENTED();
+}
+
+void
+setup_usermode_task_regs(regs_t *r, void *entry, void *stack_addr)
+{
+   NOT_IMPLEMENTED();
+}
 
 int setup_sig_handler(struct task *ti,
                       enum sig_state sig_state,
@@ -79,22 +89,6 @@ kthread_create2(kthread_func_ptr func, const char *name, int fl, void *arg)
 }
 
 void kthread_exit(void)
-{
-   NOT_IMPLEMENTED();
-}
-
-void
-finalize_usermode_task_setup(struct task *ti, regs_t *user_regs)
-{
-   NOT_IMPLEMENTED();
-}
-
-int setup_process(struct elf_program_info *pinfo,
-                  struct task *ti,
-                  const char *const *argv,
-                  const char *const *env,
-                  struct task **ti_ref,
-                  regs_t *r)
 {
    NOT_IMPLEMENTED();
 }
